+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Fixed sovereign transfer of NFT amount defaults to 0](https://github.com/multiversx/mx-lite-wallet-dapp/pull/66)
+
 ## [[1.0.3](https://github.com/multiversx/mx-lite-wallet-dapp/pull/65)] - 2024-11-15
 
 - [Updated packages](https://github.com/multiversx/mx-lite-wallet-dapp/pull/64)